make many apis for transactrions

+++ src/middleware/admin/multer.middleware.js
@@ -34,7 +34,7 @@ const getFolder = (fieldname, emailPrefix) => {
     return `src/images/therapists/${emailPrefix}`;
 
   switch (fieldname) {
-    case "blog":
+    case "blogImageUrl":
       return "src/images/blogs";
     case "userAvatar":
       return "src/images/users";
@@ -43,12 +43,21 @@ const getFolder = (fieldname, emailPrefix) => {
   }
 };
 
-// Function to remove existing file by fieldname
-const removeExistingFile = (folder, fieldname) => {
+// Function to remove existing file by fieldname and filename for blogImageUrl
+const removeExistingFile = (folder, fieldname, req) => {
   try {
     const files = fs.readdirSync(folder);
     for (const file of files) {
       if (file.includes(fieldname)) {
+        if (fieldname === "blogImageUrl") {
+          const titlePath = req.body?.title
+            .toLowerCase()
+            .split(" ")
+            .slice(0, 3)
+            .join("-");
+          const isBlogImage = file.includes(titlePath);
+          if (!isBlogImage) continue;
+        }
         const filePath = path.join(folder, file);
         fs.unlinkSync(filePath);
         console.log(`Removed old file: ${filePath}`);
@@ -69,7 +78,7 @@ const storage = multer.diskStorage({
 
     try {
       createFolder(folder); // Ensure folder exists
-      removeExistingFile(folder, file.fieldname); // Remove existing file with same fieldname
+      removeExistingFile(folder, file.fieldname, req); // Remove existing file with same fieldname
       cb(null, folder);
     } catch (error) {
       cb(new Error("Failed to create directory"), folder); // Handle any folder creation errors
@@ -78,6 +87,15 @@ const storage = multer.diskStorage({
 
   filename: function (req, file, cb) {
     const ext = path.extname(file.originalname);
+    if (file.fieldname === "blogImageUrl") {
+      const createPath = req.body?.title
+        .toLowerCase()
+        .split(" ")
+        .slice(0, 3)
+        .join("-");
+      const uniqueName = `${Date.now()}-${file.fieldname}${createPath}${ext}`; // Generate unique file name
+      return cb(null, uniqueName);
+    }
     const uniqueName = `${Date.now()}-${file.fieldname}${ext}`; // Generate unique file name
     cb(null, uniqueName);
   },
@@ -89,13 +107,8 @@ const fileFilter = (req, file, cb) => {
     allowedTypes.test(path.extname(file.originalname).toLowerCase()) &&
     allowedTypes.test(file.mimetype);
 
-  if (isValid) {
-    cb(null, true);
-  } else {
-    cb(
-      new ApiError(400, "", "Only JPG, WEBP, PNG images, and PDFs are allowed")
-    );
-  }
+  if (isValid) cb(null, true);
+  else cb(new ApiError(400, "", "Only JPG, WEBP, PNG are allowed"));
 };
 
 const upload = multer({ storage: storage, fileFilter: fileFilter });
